otel: refactor tracing helpers for child processes
Merge the `tracing/env` and `tracing/exec` package into a single package
with some common types shared between the two.
