+++ util/tracing/childprocess/traceexec.go
@@ -1,14 +1,9 @@
-package detect
+package childprocess
 
 import (
 	"context"
 
 	"go.opentelemetry.io/otel/propagation"
-)
-
-const (
-	traceparentHeader = "traceparent"
-	tracestateHeader  = "tracestate"
 )
 
 // Environ returns list of environment variables that need to be sent to the child process
@@ -38,40 +33,4 @@ func Environ(ctx context.Context) []string {
 	}
 
 	return env
-}
-
-type textMap struct {
-	parent string
-	state  string
-}
-
-func (tm *textMap) Get(key string) string {
-	switch key {
-	case traceparentHeader:
-		return tm.parent
-	case tracestateHeader:
-		return tm.state
-	default:
-		return ""
-	}
-}
-
-func (tm *textMap) Set(key string, value string) {
-	switch key {
-	case traceparentHeader:
-		tm.parent = value
-	case tracestateHeader:
-		tm.state = value
-	}
-}
-
-func (tm *textMap) Keys() []string {
-	var k []string
-	if tm.parent != "" {
-		k = append(k, traceparentHeader)
-	}
-	if tm.state != "" {
-		k = append(k, tracestateHeader)
-	}
-	return k
 }
